#121 Code cleanup
