+++ extension/ts/main/adapters/tags/types.ts
@@ -1,0 +1,8 @@
+type TagNode = {tag: string; parent?: TagNode};
+type TagTree = Map<string, TagNode>;
+
+interface TagSearchOptions {
+	noCache: boolean;
+}
+
+export {TagNode, TagTree, TagSearchOptions};
